Event Handling and Signaling in Linux
• Demonstrated imprecise computation programming pattern in which
executing computation will be terminated, using setjmp & longjmp and the
existing results are put out on double click of right mouse button.
• Illustrated Linux's Signal handling facility for various thread states
(running, runnable, blocked by a semaphore, delayed). Used “trace_cmd”
to collect events from the Linux internal tracer ftrace. The trace
records were then analyzed via a GUI front end kernelshark to find out
time the signal handler was generated and time at which it was
delivered.
• Developed a Library routine, for POSIX threads to register itself to
receive the signal to be delivered to it. Threads not registered to the
Library did not receive the generated signal.